Revert "Show summary of work hours per day in a given range (#16)"
This reverts commit c4fa948644b323d11c471c65c3997f3ad2705d2d.
