+++ requirements.txt
@@ -3,4 +3,3 @@ pandas~=1.3.5
 coverage~=6.2
 coverage-badge~=1.1.0
 num2words~=0.5.10
-numpy~=1.21.0
