--- src/commands/status/print_list_item.py
@@ -1,6 +1,0 @@
-from src.console_logger.console_logger import info
-
-
-def print_list_item(message: str) -> None:
-    style = " - "
-    info(style + message)
