--- src/exceptions/InvalidIDOfDateException.py
@@ -1,2 +1,0 @@
-class InvalidIDOfDateException(Exception):
-    pass
