--- src/handler/entry_handler/GroupedEntryHandler.py
@@ -1,52 +1,0 @@
-import pandas
-
-from src.handler.entry_handler.BaseEntryHandlerClass import BaseEntryHandlerClass
-
-
-class GroupedEntryHandler(BaseEntryHandlerClass):
-    def __init__(self):
-        super(GroupedEntryHandler, self).__init__()
-
-    def get_entries_grouped_by_date(self) -> pandas.DataFrame:
-        self.data = self.get_data()
-
-        self.data = self.group_entries_by_date()
-        self.data = self.sum_work_hours_up()
-        self.data = self.rename_columns()
-
-        self.change_index()
-
-        self.data = self.reorder_entries_from_the_latest_down_to_the_oldest_one()
-
-        return self.data
-
-    def sum_work_hours_up(self) -> pandas.DataFrame:
-        data_with_sum_of_work_hours = self.data \
-            .agg(
-                    work_hours=pandas.NamedAgg(column="work_hours", aggfunc="sum"),
-                    individual_entries=pandas.NamedAgg(column="start_date", aggfunc="size")
-                )
-        return data_with_sum_of_work_hours
-
-    def rename_columns(self) -> pandas.DataFrame:
-        data_with_renamed_columns = self.data \
-            .rename(columns={
-                                "start_date": "Date", "work_hours": "Total work time",
-                                "individual_entries": "Individual entries"
-                            }
-                    )
-        return data_with_renamed_columns
-
-    def change_index(self) -> None:
-        self.name_index_column()
-        self.boost_index()
-
-    def name_index_column(self) -> None:
-        self.data.index.name = "ID"
-
-    def boost_index(self) -> None:
-        self.data.index += 1
-
-    def reorder_entries_from_the_latest_down_to_the_oldest_one(self) -> pandas.DataFrame:
-        reordered_data = self.data.iloc[::-1]
-        return reordered_data
