+++ tests/test_click_commands/CommandBaseTestingClass.py
@@ -4,9 +4,7 @@ import os
 from click.testing import CliRunner
 
 from src.main import cli, add_subcommands_to_cli
-
 from src.csv.CSVHandler import CSVHandler
-from src.setup_test_values.setup_test_values import SetupTestValues
 
 
 class CommandBaseTestingClass(unittest.TestCase):
@@ -15,7 +13,6 @@ class CommandBaseTestingClass(unittest.TestCase):
         self.result = None
 
         self.csv_handler = CSVHandler()
-        self.set_upper = SetupTestValues()
 
         add_subcommands_to_cli()
 
@@ -32,7 +29,3 @@ class CommandBaseTestingClass(unittest.TestCase):
         except FileNotFoundError:
             # File is not existing, that's good, nothing to do for us.
             pass
-
-    def setup_test_values(self, number_of_entries: int = 4) -> None:
-        self.set_upper.set_number_of_entries(number_of_entries)
-        self.set_upper.setup()
